corrected type hints with mypy
- added hyperparameter to MLPC and SVC

+++ sam_ml/models/BaggingClassifier.py
@@ -47,7 +47,7 @@ class BC(Classifier):
         base_estimator = [DecisionTreeClassifier(max_depth=i) for i in range(1,11)]+[SVC(probability=True, kernel='linear'), LogisticRegression(), GradientBoostingClassifier(), RandomForestClassifier(max_depth=5), KNeighborsClassifier()],
         n_estimators: list[int] = list(range(10, 101, 10)) + [3, 4, 5, 6, 7, 8, 9, 200, 500, 1000, 1500, 3000],
         max_samples: list[float] = [0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1],
-        max_features: list[int] = [0.5, 1.0, 2, 4],
+        max_features: list[float] = [0.5, 1.0, 2, 4],
         bootstrap: list[bool] = [True, False],
         bootstrap_features: list[bool] = [True, False],
         n_split_num: int = 5,
